Add additional safety check on PeriphMode message processing.
ButteRFly crashes when no advertising records are set to an empty
buffer, so we added a safety check to make sure that the provided
advertising data (containing advertising records) contains at least
a single record in order to be taken into account. If this is not
the case, a ParameterError is returned to the host.